# Optimize formatting of strings with occasional escape sequences