@dsnet 2025-01-27
> I'm a bit uneasy with slight changes causing the compiler to give better performance when the code should be otherwise equivalent. Should we file a bug with a short example? LGTM in any case.

When I think about the code and how it matches what the CPU actually executes, the more optimal logical flow makes sense why it'd be faster. That said, I'm not sure the compiler can do this optimization for you unless it can infer which branch is more hotly taken (or not taken). This is possible with PGO, but sounds like a hard problem without guidance.